Fix d/watch
Unfortunately, the upstream tgz has an incomplete version number
embedded in its name, so we need to be a bit creative.

pagemangle extracts the correct version from the link text and adds it
as bogus parameter "?version=..." to the download URL. Then, uscan can
match on that bogus parameter to learn the upstream version number, and
downloadurlmangle removes it again for actual download.

It's not pretty, but it gets the job done.